Move federation to plugin-helpers
