+++ packages/utils/plugins-helpers/src/index.ts
@@ -3,3 +3,4 @@ export { toPascalCase } from './to-pascal-case';
 export * from './types';
 export * from './utils';
 export * from './helpers';
+export * from './federation';
