+++ packages/utils/plugins-helpers/src/utils.ts
@@ -1,3 +1,4 @@
+import { GraphQLOutputType, GraphQLNamedType, GraphQLNonNull, GraphQLList, isListType, isNonNullType } from 'graphql';
 import { Types } from './types';
 
 export function mergeOutputs(content: Types.PluginOutput | Array<Types.PluginOutput>): string {
@@ -16,4 +17,16 @@ export function mergeOutputs(content: Types.PluginOutput | Array<Types.PluginOut
   }
 
   return [...result.prepend, result.content, ...result.append].join('\n');
+}
+
+export function isWrapperType(t: GraphQLOutputType): t is GraphQLNonNull<any> | GraphQLList<any> {
+  return isListType(t) || isNonNullType(t);
+}
+
+export function getBaseType(type: GraphQLOutputType): GraphQLNamedType {
+  if (isWrapperType(type)) {
+    return getBaseType(type.ofType);
+  } else {
+    return type;
+  }
 }
